+++ projects/iad-interface-admin/src/lib/iad-base-grid/services/iad-base-grid-actions-service.ts
@@ -6,7 +6,8 @@ export enum IAD_GRID_ACTIONS {
     CLEAR = 'clear',
     UNSELECT = 'unselect',
     RESET_FILTER = 'resetFilter',
-    RESET_SCROLLABLE_TABLE_WIDTH = 'resetScrollableTableWidth'
+    RESET_SCROLLABLE_TABLE_WIDTH = 'resetScrollableTableWidth',
+    RESET_GRID = 'resetGrid'
 }
 
 export interface IadBaseGridAction {
@@ -20,23 +21,54 @@ export class IadBaseGridActionsService {
 
     action: Subject<IadBaseGridAction> = new Subject();
 
+    /**
+     * Add search value to global search
+     * @param gridId
+     * @param value
+     */
     doGlobalSearch(gridId, value) {
         this.action.next({gridId, value, action: IAD_GRID_ACTIONS.GLOBAL_SEARCH});
     }
 
+    /**
+     * Clears all filters and refresh table
+     * @param gridId
+     */
     doClear(gridId) {
         this.action.next({gridId, action: IAD_GRID_ACTIONS.CLEAR});
     }
 
+    /**
+     * Unselect currently selected row in grid
+     * @param gridId
+     * @param value
+     */
     doUnselect(gridId, value) {
         this.action.next({gridId, value, action: IAD_GRID_ACTIONS.UNSELECT});
     }
 
+    /**
+     * Reset table data filter for particular column
+     * @param gridId
+     * @param value
+     */
     doResetFilter(gridId, value) {
         this.action.next({gridId, value, action: IAD_GRID_ACTIONS.RESET_FILTER});
     }
 
+    /**
+     * Reset scrollable table width to its calculated value
+     * @param gridId
+     */
     doResetScrollableTableWidth(gridId) {
         this.action.next({gridId, action: IAD_GRID_ACTIONS.RESET_SCROLLABLE_TABLE_WIDTH});
+    }
+
+    /**
+     * Reset grid settings and grid width
+     * @param gridId
+     */
+    doResetGrid(gridId) {
+        this.action.next({gridId, action: IAD_GRID_ACTIONS.RESET_GRID});
     }
 }